Fixed duplicate bookmarks bug
git-svn-id: https://svn2.assembla.com/svn/savant/Savant/trunk@1408 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ src/savant/util/MiscUtils.java
@@ -66,11 +66,11 @@ public class MiscUtils {
      * @param num The number to format
      * @return A formatted string
      */
-     public static String numToString(float num) {
+     public static String numToString(double num) {
          return numToString(num,0);
      }
 
-    public static String numToString(float num, int significantdigits) {
+    public static String numToString(double num, int significantdigits) {
         //TODO: implement formatter
         String formatString = "###,###";
